+++ CHANGELOG.md
@@ -50,6 +50,7 @@
   - `XargsDirname()` now supports tracing
   - `XargsRmFile()` now supports tracing
   - `XargsTestFilepathExists()` now supports tracing
+  - `XargsTruncateFiles()` now supports tracing
 * Pipelines now set a context flag in their Pipe
   - this is used to tell sinks where to read from!